+++ algebra/src/reduce/macros.rs
